Gui: extract TranslateManip from label nodes
Move TranslateManip into dedicated source and header files, update its view-provider user to include the new declaration directly, and keep node registration explicit in SoFCDB. This separates the transform manipulator from the unrelated label node collection without changing behavior.
