+++ src/Gui/ViewProviderAnnotation.h
@@ -27,7 +27,6 @@
 #include <App/PropertyUnits.h>
 #include <Base/Vector3D.h>
 #include <optional>
-#include "SoTextLabel.h"
 
 class SoFont;
 class SoText2;
@@ -41,6 +40,8 @@ class SoCoordinate3;
 
 namespace Gui
 {
+
+class TranslateManip;
 
 class GuiExport ViewProviderAnnotation: public ViewProviderDocumentObject
 {
